Add files via upload
Hough Transform:

There are two planes seen, a line y = a’x + b’y  in x-y plane is set of points those represent a point in a’ - b’ plane. A point in xy-plane may have many lines go through it, which is mapped into a line in ab-plane.  And hence we can use hough transform to draw edge detection by choosing a point in a’-b’ space where maximum lines converge. This point must be representing a line since all those points will satisfy the equation of a single line. To get the points in hough space, we use following heisenberg’s formula to represent line:
Preprocessing: using sobel edge detector to detect the edges in the image

For each image pixel, for wide range of theta(ranging from -90 to 90), calculate rho such as:
Rho = x(cos (theta)) + y(sin(theta))

Create a vote matrix where the rows of the vote matrix are values of rhos and columns of the vote matrix are values of thetas, each cell in the vote matrix will represent the count for which the same rho and theta values clashed. 
	The dimension of the vote matrix will be (the length of rhos) x (values of thetas)
	If we closely observe rho function, we understand that it calculates sin and cos values for same theta, this results in sin wave (as proven in Homework 3, Q. 3.b) Hence the value of cos (theta) will be increasing while the value of sin(theta) is reducing and vice versa. Only for tehta 45, the value will be maximum. And it will be 1/sqrt(2). Hence we consider the length of rho to be maximum of 2 x (len(diagonal of the image)).

Select a threshold, chose the values of rho and theta for which, the votes were greater than the threshold. 
Convert such values of rho and theta into x - y coordinates by following steps:
For each corresponding values of theta and rhos, 
X0 = rho * cos(theta)
Y0 = rho * sin(theta)
Now that we have the values of x0 and y0 for all theta and rho values, we calculate the end points of the line as:
x1 = (x0 + 1000*(- sin(theta)))
y1 = (y0 + 1000*(cos(theta)))
x2 = (x0 - 1000*(-sin(theta)))
y2 = (y0 - 1000*(cos (theta)))

Since sin(-(theta)) = -sin(theta) and cos(-(theta)) = cos(theta)

Plot them on the original image, they will return the detected lines. 
	Now that we two endpoints of the line named (x0,y0), (x1,y1) we can plot the line using them.
The blue lines can be detected by applying a threshold of less than 3.10 theta. Since we can observe the value of thetas and we see that 3.10 is the most repeated pattern. Considering theta to be the slope, we can differentiate red and blue lines since the angle of them is 0 and 45 degree respectively. 
Multiple lines are observed on the same line since the edges detected are broad.